pyproject.toml
[tool.poetry]
name = "reddit-scrapper"
version = "0.1.0"
description = ""
authors = ["David <david@dzhy.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
praw = "^7.2.0"
pandas = "^1.2.3"
python-dotenv = "^0.16.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pylint = "^2.7.3"
yapf = "^0.31.0"
mypy = "^0.812"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
